requirements.txt
psycopg2-binary
pandas
requests
datetime
streamlit
altair